Split ForgeWindowUI into CoreSlots, GearSlots, and UIUpdates partials
