--- Assets/Scripts/Gear/UI/ForgeWindowUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ba45dfc14cf9ac9439f07a22fef7a587
